// src/types/terms.ts
export type TermType = {
	title: string;
	icon: string;
};
export type TermsType = {
	personnel_units: TermType;
	tanks: TermType;
	armoured_fighting_vehicles: TermType;
	artillery_systems: TermType;
	mlrs: TermType;
	aa_warfare_systems: TermType;
	planes: TermType;
	helicopters: TermType;
	vehicles_fuel_tanks: TermType;
	warships_cutters: TermType;
	uav_systems: TermType;
	special_military_equip: TermType;
	atgm_srbm_systems: TermType;
	cruise_missiles: TermType;
};